Update cacaos, rename translation modules and feats.
